requirement.txt
torch
torchvision
transformers
Pillow
datasets
nltk
numpy
matplotlib
